Implement OptiGlue interface
Change entry point to optiglue
